# docs/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Portal AEGEA - API",
    "version": "1.0.0",
    "description": "API do Portal de Produtos Digitais e Automacoes do CoE Power Platform - AEGEA",
    "contact": {
      "name": "CoE Power Platform - AEGEA"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via POST /auth/login (mock) ou Microsoft Entra ID"
      }
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@empresa.mock"
          },
          "password": {
            "type": "string",
            "example": "admin"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "role": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login (mock)",
        "description": "Mock auth. Usuarios: admin@empresa.mock, editor@empresa.mock, viewer@empresa.mock (senha: admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token JWT + dados do usuario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais invalidas"
          }
        }
      }
    },
...
  }
}